Strongly-typed GraphQL queries (#12)
