+++ src/api/common/types.ts
@@ -1,0 +1,4 @@
+export enum OrderDirection {
+  Ascending = 'ASC',
+  Descending = 'DESC',
+}
