+++ src/api/profile/useProfile.ts
@@ -1,0 +1,5 @@
+import { GetProfile, GetProfileVariables } from './GetProfile'
+import { useQuery } from '@apollo/client'
+
+export const useProfile = (variables: GetProfileVariables) =>
+  useQuery<GetProfile, GetProfileVariables>(GetProfile, { variables })
